Modularize parsing of args + building of cap.
